+++ Types/DAT_v06.cs
@@ -12,7 +12,7 @@ namespace BrickVault.Types
     {
         public override uint Version() => 6;
 
-        public DAT_v06(RawFile file, uint trailerOffset, uint trailerSize) : base(file, trailerOffset, trailerSize)
+        public DAT_v06(RawFile file, long trailerOffset, uint trailerSize) : base(file, trailerOffset, trailerSize)
         {
 
         }
