+++ Types/DAT_v08.cs
@@ -10,7 +10,7 @@ namespace BrickVault.Types
     {
         public override uint Version() => 8;
 
-        public DAT_v08(RawFile file, uint trailerOffset, uint trailerSize) : base(file, trailerOffset, trailerSize)
+        public DAT_v08(RawFile file, long trailerOffset, uint trailerSize) : base(file, trailerOffset, trailerSize)
         {
 
         }
