+++ Types/DAT_v11.cs
@@ -10,7 +10,7 @@ namespace BrickVault.Types
     {
         public override uint Version() => 11;
 
-        public DAT_v11(RawFile file, uint trailerOffset, uint trailerSize) : base(file, trailerOffset, trailerSize)
+        public DAT_v11(RawFile file, long trailerOffset, uint trailerSize) : base(file, trailerOffset, trailerSize)
         {
 
         }
